refactor(import): changing the way of calling components to one that occupies fewer lines
