+++ src/components/index.ts
@@ -1,0 +1,5 @@
+export { Header } from "./Header";
+export { Card } from "./Card";
+export { CardPDF } from "./Pdf";
+export { Search } from "./Search";
+export { Footer } from "./Footer";
